correção de erros - final!

+++ Grupo2A/Model/DTO/Ingrediente2listing_dto.cs
@@ -2,7 +2,7 @@ namespace Cozinha_BE.Model.DTO;
 
 public class Ingrediente2listing_dto
 {
-    public int IdIngrediente { get; set; }
+    public long IdIngrediente { get; set; }
     public required string Nome { get; set; }
     public required string Categoria { get; set; }
 }
